requirement.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

python-jose[cryptography]
passlib[bcrypt]

python-dotenv

pydantic
email-validator

fastapi-mail
